requirements.txt
ipykernel
langchain-openai
langchain
langchain_community
langchain-ollama
python-dotenv
langchain_core
langchain-groq
streamlit
faiss-cpu
langchain_chroma
pypdf
langchain-text-splitters
chromadb
langchain_huggingface
streamlit
bs4